// apps/playground-flow-refactor/src/app/components/workflow/types.ts
export enum WorkflowVersionFilterOptions {
  all = "all",
  onlyYours = "onlyYours",
}

export enum VersionHistoryContextMenuOptions {
  restore = "restore",
  edit = "edit",
  delete = "delete",
}